requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
platformdirs==4.4.0
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
urllib3==2.5.0
openai==1.108.0
PyYAML==6.0.2
inquirer==3.4.0
prompt-toolkit==3.0.52
textual>=0.47.0
sseclient-py==1.8.0
